src/leetcode/arrays_hashing/leetcode217_contains_duplicate/Solution.java
package leetcode.arrays_hashing.leetcode217_contains_duplicate;

import java.util.Arrays;
import java.util.HashSet;
import java.util.Set;

public class Solution {

    /**
     * Time : O(n)
     * Space: O(n)
     */
    public boolean containsDuplicate(int[] nums) {
        Set<Integer> set = new HashSet<>();
        for (int num : nums) {
            if (set.contains(num)) return true;
            set.add(num);
        }
        return false;
    }

    /**
     * Time : O(n log n)
     * Space: O(1)
     */
    public boolean containsDuplicate1(int[] nums) {
        Arrays.sort(nums);
        for (int i = 1; i < nums.length; i++) {
            if (nums[i - 1] == nums[i]) return true;
        }

        return false;
    }
}
